+++ src/controllers/SubscriptionController.php
@@ -204,7 +204,25 @@ class SubscriptionController extends Controller
             && !$molliePayment->subscriptionId
         ) {
             MolliePayments::$plugin->mollie->createSubscription($subscriptionElement);
-            return;
+        }
+    }
+
+
+    public function actionCheckTransactionStatus($id, $redirect)
+    {
+        try {
+            $transaction = MolliePayments::getInstance()->transaction->getTransactionbyId($id);
+            $element = Subscription::findOne(['id' => $transaction->payment]);
+            $form = MolliePayments::getInstance()->forms->getFormByid($element->formId);
+            $molliePayment = MolliePayments::getInstance()->mollie->getStatus($id, $form->handle);
+
+            if ($transaction->status !== $molliePayment->status) {
+                MolliePayments::getInstance()->transaction->updateTransaction($transaction, $molliePayment);
+                return $this->asSuccess("Transaction status updated", [], $redirect);
+            }
+            return $this->asSuccess("Transaction already up to date", [], $redirect);
+        } catch (\Throwable $e) {
+            return $this->asFailure("Something went wrong checking the status for this payment", [], $redirect);
         }
     }
 
@@ -284,7 +302,7 @@ class SubscriptionController extends Controller
         $page = $this->request->getQueryParam('page', 1);
         $baseUrl = 'mollie-payments/subscription/get-subscribers';
         $data = MolliePayments::getInstance()->subscriber->getAllSubscribers();
-        $subscribers = collect($data)->map(function ($subscriber) {
+        $subscribers = collect($data)->map(function($subscriber) {
             return [
                 'title' => $subscriber->email,
                 'id' => $subscriber->customerId,
@@ -340,7 +358,7 @@ class SubscriptionController extends Controller
             MolliePayments::getInstance()->mollie->deleteCustomer($this->request->getRequiredBodyParam('id'));
             MolliePayments::getInstance()->subscriber->deleteById($this->request->getRequiredBodyParam('id'));
             return $this->asJson(['success' => true]);
-        } catch(\Throwable $e) {
+        } catch (\Throwable $e) {
             Craft::error($e->getMessage(), MolliePayments::class);
             return $this->asJson(['success' => false]);
         }